small fix to prevent errors with NaN values in check_value()
- Integers cannot be NaN, and so "NaN" strings in the nodes textbox would cause errors when check_value() was called. As there is no need for these values to be int rather than float this small fix resolves the problem. NaN values do not impose limits on the number of Q bins.

+++ src/sas/qtgui/Calculators/GenericScatteringCalculator.py
@@ -671,9 +671,9 @@ class GenericScatteringCalculator(QtWidgets.QDialog, Ui_GenericScatteringCalcula
             else:
                 self.txtQxMax.setStyleSheet('background-color: rgb(255, 255, 255);')
         if (text_edit == self.txtNoQBins or update_all) and self.txtNoQBins.text():
-            xnodes = int(self.txtXnodes.text())
-            ynodes = int(self.txtYnodes.text())
-            znodes = int(self.txtZnodes.text())
+            xnodes = float(self.txtXnodes.text())
+            ynodes = float(self.txtYnodes.text())
+            znodes = float(self.txtZnodes.text())
             value = float(str(self.txtNoQBins.text()))
             max_step =  3*max(xnodes, ynodes, znodes) 
                 #limits qmin > maxq / nodes
